pyproject.toml
[project]
name = "reconhecimento-facial-com-deepface"
version = "0.1.0"
description = ""
authors = [
    {name = "Kauan Rodrigues",email = "kauanrl09@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "deepface (>=0.0.95,<0.0.96)",
    "tensorflow (>=2.20.0,<2.21.0)",
    "tf-keras (>=2.20.1,<3.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "numpy (>=2.1.0,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
